requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi-cli==0.0.6
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.56.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
slack_bolt==1.21.2
slack_sdk==3.33.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
